% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_half_dotplot.R, R/geom_half_boxplot.R,
%   R/geom_half_point.R, R/geom_half_point_panel.R, R/geom_half_violin.R,
%   R/stat-half-bindot.R, R/stat-half-point.R, R/stat-half-ydensity.R,
%   R/ggproto-classes.R
\name{GeomHalfDotplot}
\alias{GeomHalfDotplot}
\alias{GeomHalfBoxplot}
\alias{GeomHalfPoint}
\alias{GeomHalfPointPanel}
\alias{GeomHalfViolin}
\alias{StatHalfBindot}
\alias{StatHalfPoint}
\alias{StatHalfYdensity}
\alias{gghalves-extensions}
\title{gghalves extensions to ggplot2}
\description{
gghalves extensions to ggplot2
}
