% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghalves.R
\docType{package}
\name{gghalves-package}
\alias{gghalves}
\alias{gghalves-package}
\title{gghalves: Compose half-half plots using your favourite geoms.}
\description{
A 'ggplot2' extension for easy plotting of half-half geom combinations. Think half boxplot and half jitterplot, or half violinplot and half dotplot.
}
\details{
D1
D2
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erocoar/gghalves}
}

}
\author{
\strong{Maintainer}: Frederik Tiedemann \email{fj.tiedemann@googlemail.com}

}
\keyword{internal}
