% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-half-dotplot.R, R/geom_half_boxplot.R,
%   R/geom_half_point.R, R/geom_half_violin.R, R/stat-half-bindot.R,
%   R/stat-half-point.R, R/stat-half-ydensity.R, R/ggproto-classes.R
\docType{data}
\name{GeomHalfDotplot}
\alias{GeomHalfDotplot}
\alias{GeomHalfBoxplot}
\alias{GeomHalfPoint}
\alias{GeomHalfViolin}
\alias{StatHalfBindot}
\alias{StatHalfPoint}
\alias{StatHalfYdensity}
\alias{gghalves-extensions}
\title{ggpol extensions to ggplot2}
\description{
ggpol extensions to ggplot2
}
\keyword{datasets}
